Cleanup Cygwin builds. * windows/Makefile.am (loaddrv_LDFLAGS): remove, the -mno-cygwin flag is supposed to be set in CFLAGS by ./configure * configure.ac: add a check for the presence of usleep(), add a check whether the linker accepts -static * avrdude.h: protect prototype for usleep by !defined(HAVE_USLEEP) * ppwin.c (usleep): protect by !defined(HAVE_USLEEP) * main.c: silence "array subscript of type char" compiler warnings by casting all arguments to tolower()/toupper() and isspace()/ isdigit()/ispunct() to "int" * butterfly.c: (Dito.) * avr910.c: (Dito.)
git-svn-id: svn://svn.savannah.nongnu.org/avrdude/trunk@936 81a1dc3b-b13d-400b-aceb-764788c761c2

+++ avrdude/avr910.c
@@ -624,7 +624,7 @@ static int avr910_paged_write(PROGRAMMER * pgm, AVRPART * p, AVRMEM * m,
     cmd = malloc(4 + blocksize);
     if (!cmd) rval = -1;
     cmd[0] = 'B';
-    cmd[3] = toupper(m->desc[0]);
+    cmd[3] = toupper((int)(m->desc[0]));
 
     while (addr < max_addr) {
       if ((max_addr - addr) < blocksize) {
@@ -714,7 +714,7 @@ static int avr910_paged_load(PROGRAMMER * pgm, AVRPART * p, AVRMEM * m,
     int blocksize = PDATA(pgm)->buffersize;
 
     cmd[0] = 'g';
-    cmd[3] = toupper(m->desc[0]);
+    cmd[3] = toupper((int)(m->desc[0]));
 
     avr910_set_addr(pgm, addr);
